:tractor: :newspaper: Splitting cli into its own module
There were too many functions for the CLI in just one file with
more to come.
This change now splits all of them into individual, hidden files.
