+++ deepblink/__main__.py
@@ -1,13 +1,5 @@
-"""Entrypoint module, in case you use `python -mdeepblink`.
+"""Entrypoint module, in case you use `python -m deepblink`."""
 
-Read the docs in the usage section or for the `cli` module for information.
-
-Why does this file exist, and why __main__? For more info, read:
-
-- https://www.python.org/dev/peps/pep-0338/
-- https://docs.python.org/2/using/cmdline.html#cmdoption-m
-- https://docs.python.org/3/using/cmdline.html#cmdoption-m
-"""
 from .cli import main
 
 if __name__ == "__main__":
