+++ deepblink/cli/_handler.py
@@ -1,31 +1,8 @@
-"""Module that contains the command line app.
+"""Event handlers for deepblinks CLI."""
 
-Used for inferencing of new data with pretrained models.
-Because the model is loaded into memory for every run, it is faster to
-run multiple images at once by passing a directory as input.
-
-Usage:
-    ``$ deepblink [-h] [-o OUTPUT] [-t {csv,txt}] [-r RADIUS] [-v] [-V] MODEL INPUT``
-    ``$ deepblink --help``
-
-Positional Arguments:
-    MODEL           model .h5 file location
-    INPUT           input file/folder location
-
-Optional Arguments:
-    -h, --help      show this help screen
-
-    -o, --output    output file/folder location [default: input location]
-    -t, --type      output file type [options: csv, txt] [default: csv]
-    -r, --radius    if given, calculate the integrated intensity in the given radius around each coordinate
-    -v, --verbose   set program output to verbose [default: quiet]
-    -V, --version   show version program's version number and exit
-"""
 from typing import List
-import argparse
 import logging
 import os
-import sys
 import textwrap
 import yaml
 
@@ -33,196 +10,16 @@ import numpy as np
 import skimage.io
 import tensorflow as tf
 
-from .inference import get_intensities
-from .inference import predict
-from .io import basename
-from .io import grab_files
-from .io import load_image
-from .io import load_model
-from .training import run_experiment
-
-# Removes tensorflow's information on CPU / GPU availablity.
-os.environ["TF_CPP_MIN_LOG_LEVEL"] = "2"
+from ..inference import get_intensities
+from ..inference import predict
+from ..io import basename
+from ..io import grab_files
+from ..io import load_image
+from ..io import load_model
+from ..training import run_experiment
 
 # List of currently supported image file extensions.
 EXTENSIONS = ["tif", "jpeg", "jpg", "png"]
-
-
-class FileFolderType:
-    """Custom type supporting folders or files."""
-
-    def __init__(self):
-        pass
-
-    def __call__(self, value):  # noqa: D102
-        """Python type internal function called by argparse to check input."""
-        if not any((os.path.isdir(value), os.path.isfile(value))):
-            raise argparse.ArgumentTypeError(
-                f"Input value must be file or folder. '{value}' is not."
-            )
-        return value
-
-
-class FolderType:
-    """Custom type supporting folders."""
-
-    def __init__(self):
-        pass
-
-    def __call__(self, value):  # noqa: D102
-        """Python type internal function called by argparse to check input."""
-        if not os.path.isdir(value):
-            raise argparse.ArgumentTypeError(
-                f"Input value must be folder and must exist. '{value}' is not."
-            )
-        return value
-
-
-def _parse_args_train(
-    subparsers: argparse._SubParsersAction,
-) -> argparse._SubParsersAction:
-    """Subparser for training."""
-    parser = subparsers.add_parser(
-        "train", help="\U0001F35E train a freshly baked model on a dataset",
-    )
-    parser.add_argument(
-        "-c",
-        "--config",
-        type=str,
-        required=True,
-        help="path to the experimental config.yaml file. Check the GitHub repository for an example",
-    )
-    parser.add_argument(
-        "-g",
-        "--gpu",
-        type=int,
-        default=None,
-        help="index of GPU to be used [default: None]",
-    )
-    return subparsers
-
-
-def _parse_args_check(
-    subparsers: argparse._SubParsersAction,
-) -> argparse._SubParsersAction:
-    """Subparser for checking."""
-    parser = subparsers.add_parser(
-        "check", help="\U0001F537 \U0001F535 determine your input images' shape"
-    )
-    parser.add_argument(
-        "INPUT", type=str, help=f"input image location [filetypes: {EXTENSIONS}]",
-    )
-    return subparsers
-
-
-def _parse_args_predict(
-    subparsers: argparse._SubParsersAction,
-) -> argparse._SubParsersAction:
-    """Subparser for prediction."""
-    parser = subparsers.add_parser(
-        "predict",
-        help="\U0001F914 inference / prediction of data with a pre-trained model",
-    )
-    parser.add_argument(
-        "MODEL", type=argparse.FileType("r"), help="model .h5 file location"
-    )
-    parser.add_argument(
-        "INPUT",
-        type=FileFolderType(),
-        help=f"input file/folder location [filetypes: {EXTENSIONS}]",
-    )
-    parser.add_argument(
-        "-o",
-        "--output",
-        type=FolderType(),
-        help="output file/folder location [default: input location]",
-    )
-    parser.add_argument(
-        "-t",
-        "--type",
-        type=str,
-        choices=["csv", "txt"],
-        default="csv",
-        help="output file type [options: csv, txt] [default: csv]",
-    )
-    parser.add_argument(
-        "-r",
-        "--radius",
-        type=int,
-        default=None,
-        help=textwrap.dedent(
-            """if given, calculate the integrated intensity
-        in the given radius around each coordinate. set radius to zero if only the
-        central pixels intensity should be calculated."""
-        ),
-    )
-    return subparsers
-
-
-def _parse_args_eval(
-    subparsers: argparse._SubParsersAction,
-) -> argparse._SubParsersAction:
-    """Subparser for evaluation."""
-    parser = subparsers.add_parser(
-        "eval", help="\U0001F3AD measure a models performance on a dataset"
-    )
-    parser.add_argument(
-        "INPUT",
-        type=FileFolderType(),
-        help=f"input file/folder location [filetypes: {EXTENSIONS}]",
-    )
-    return subparsers
-
-
-def _parse_args():
-    """Argument parser."""
-    parser = argparse.ArgumentParser(
-        prog="deepblink",
-        description=textwrap.dedent(
-            """deepBlinks command line interface \U0001F469\U0000200D\U0001F4BB
-        for training, inferencing, and evaluation"""
-        ),
-        epilog="We hope you enjoy using deepBlink \U0001F603",
-    )
-    parser.add_argument("-V", "--version", action="version", version="%(prog)s 0.0.6")
-    parser.add_argument(
-        "-v",
-        "--verbose",
-        action="store_true",
-        help="set program output to verbose [default: quiet]",
-    )
-    parser.add_argument("--debug", action="store_true", help=argparse.SUPPRESS)
-
-    subparsers = parser.add_subparsers(dest="command")
-    subparsers = _parse_args_train(subparsers)
-    subparsers = _parse_args_check(subparsers)
-    subparsers = _parse_args_predict(subparsers)
-    subparsers = _parse_args_eval(subparsers)
-
-    args = parser.parse_args()
-    return args
-
-
-def _configure_logger(verbose: bool, debug: bool):
-    """Return verbose logger with three levels.
-
-    * Verbose false and debug false - no verbose logging.
-    * Verbose true and debug false - only info level loginfo for standard users.
-    * Debug true - debug mode for developers.
-    """
-    if debug:
-        level = logging.DEBUG
-    else:
-        if verbose:
-            level = logging.INFO
-        else:
-            level = logging.CRITICAL
-
-    logging.basicConfig(
-        format="%(asctime)s: %(message)s", stream=sys.stdout, level=level
-    )
-    logger = logging.getLogger("Verbose output logger")
-    return logger
 
 
 class HandleTrain:
@@ -456,28 +253,3 @@ class HandlePredict:
             self.predict_single(fname_in, model)
 
         self.logger.log(20, "\U0001F3C1\U0001F3C3 all predictions are complete")
-
-
-def main():
-    """Entrypoint for the CLI."""
-    args = _parse_args()
-    logger = _configure_logger(args.verbose, args.debug)
-
-    if args.command == "train":
-        handler = HandleTrain(arg_config=args.config, arg_gpu=args.gpu, logger=logger)
-
-    if args.command == "check":
-        handler = HandleCheck(arg_input=args.INPUT, logger=logger)
-
-    if args.command == "predict":
-        handler = HandlePredict(
-            arg_model=args.MODEL.name,
-            arg_input=args.INPUT,
-            arg_output=args.output,
-            arg_radius=args.radius,
-            arg_type=args.type,
-            arg_verbose=args.verbose,
-            logger=logger,
-        )
-
-    handler.run()
