+++ deepblink/cli/_type.py
@@ -1,0 +1,34 @@
+"""Custom argparse types / classes."""
+
+import os
+import argparse
+
+
+class FileFolderType:
+    """Custom type supporting folders or files."""
+
+    def __init__(self):
+        pass
+
+    def __call__(self, value):  # noqa: D102
+        """Python type internal function called by argparse to check input."""
+        if not any((os.path.isdir(value), os.path.isfile(value))):
+            raise argparse.ArgumentTypeError(
+                f"Input value must be file or folder. '{value}' is not."
+            )
+        return value
+
+
+class FolderType:
+    """Custom type supporting folders."""
+
+    def __init__(self):
+        pass
+
+    def __call__(self, value):  # noqa: D102
+        """Python type internal function called by argparse to check input."""
+        if not os.path.isdir(value):
+            raise argparse.ArgumentTypeError(
+                f"Input value must be folder and must exist. '{value}' is not."
+            )
+        return value
